refactor: code frame & code stack frame error
