+++ packages/error-overlay/src/react/components/Errors/CallStackFrame/index.tsx
@@ -1,0 +1,1 @@
+export { CallStackFrame } from './CallStackFrame';
