+++ packages/error-overlay/src/react/components/Errors/CodeFrame/CodeFrame.tsx
@@ -2,7 +2,7 @@ import Anser from 'anser';
 import * as React from 'react';
 import { StackFrame } from 'stacktrace-parser';
 import stripAnsi from 'strip-ansi';
-import { getFrameSource } from '../../helpers/stack-frame';
+import { getFrameSource } from '../../../helpers/stack-frame';
 
 export type CodeFrameProps = { stackFrame: StackFrame; codeFrame: string };
