+++ packages/error-overlay/src/react/components/Errors/CodeFrame/index.tsx
@@ -1,2 +1,1 @@
 export { CodeFrame } from './CodeFrame';
-export { styles } from './styles';
